Update migrations to allow for bigger numbers

+++ database/migrations/2020_05_24_135404_create_players_table.php
@@ -17,9 +17,9 @@ class CreatePlayersTable extends Migration
             $table->id();
             $table->string('name', 100);
             $table->string('position', 3);
-            $table->tinyInteger('team')->nullable()->default(null);
+            $table->integer('team')->nullable()->default(null);
             $table->integer('college')->nullable()->default(null);
-            $table->tinyInteger('owned_by')->nullable()->default(null);
+            $table->integer('owned_by')->nullable()->default(null);
             $table->boolean('rookie')->default(false);
             $table->boolean('free_agent')->default(false);
             $table->boolean('recently_traded')->default(false);
